requirements.txt
google-genai==1.64.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
pillow==12.1.1
tqdm==4.67.3
python-dotenv==1.2.1